break index into multiple module files
