+++ lib/debug.js
@@ -1,0 +1,3 @@
+const debug = require('debug')('eslint-plugin-import:resolver:kibana');
+
+module.exports = debug;
