+++ lib/defaults.js
@@ -1,0 +1,5 @@
+module.exports = {
+  kibanaPath: '../kibana',
+  pluginDirs: [],
+  pluginPaths: [],
+};
